+++ pom.xml
@@ -59,6 +59,18 @@
       <artifactId>mybatis</artifactId>
       <version>3.5.11</version>
     </dependency>
+    <dependency>
+      <groupId>junit</groupId>
+      <artifactId>junit</artifactId>
+      <version>4.7</version>
+      <scope>compile</scope>
+    </dependency>
+    <dependency>
+      <groupId>org.projectlombok</groupId>
+      <artifactId>lombok</artifactId>
+      <version>1.18.24</version>
+    </dependency>
+
 
   </dependencies>
   <build>
