+++ src/main/java/com/example/spring2/BoardController.java
@@ -1,5 +1,7 @@
 package com.example.spring2;
 
+import com.example.spring2.board.BoardService;
+import com.example.spring2.board.BoardVO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
