--- src/main/java/com/example/spring2/UserVO.java
@@ -1,31 +1,0 @@
-package com.example.spring2;
-
-public class UserVO {
-    private String userid;
-    private String password;
-    private String username;
-
-    public String getUserid() {
-        return userid;
-    }
-
-    public void setUserid(String userid) {
-        this.userid = userid;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-}
