+++ src/main/java/com/example/spring2/board/BoardVO.java
@@ -1,0 +1,26 @@
+package com.example.spring2.board;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import java.util.Date;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class BoardVO {
+	private int seq;
+	private String title;
+	private String writer;
+	private String content;
+	private String category;
+
+	private Date regdate;
+	private String email;
+
+	private String file;
+	private int cnt;
+}
