+++ src/main/java/com/example/spring2/user/UserDAO.java
@@ -1,4 +1,4 @@
-package com.example.spring2;
+package com.example.spring2.user;
 
 import org.mybatis.spring.SqlSessionTemplate;
 import org.springframework.beans.factory.annotation.Autowired;
