+++ src/main/java/com/example/spring2/user/UserServiceImpl.java
@@ -1,4 +1,4 @@
-package com.example.spring2;
+package com.example.spring2.user;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
