+++ src/main/java/com/example/spring2/user/UserVO.java
@@ -1,0 +1,16 @@
+package com.example.spring2.user;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class UserVO {
+    private String userid;
+    private String password;
+    private String username;
+}
