+++ src/main/resources/mappers/user-mapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper namespace="User">
-    <select id="getUser" resultType="com.example.spring2.UserVO">
+    <select id="getUser" resultType="user">
     select userid, username from member
     where userid=#{userid} and password=${password}
     </select>
